requirements.txt
decord==0.6.0
einops==0.8.1
huggingface_hub==0.29.1
matplotlib==3.7.0
numpy==1.24.4
opencv_python==4.7.0.72
pyarrow==11.0.0
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.1.99
torch==2.5.1
torchvision==0.20.1
transformers==4.49.0
accelerate>=0.34.0
wandb
gradio
setuptools
wheel
ninja
bitsandbytes
xlsxwriter
triton ; sys_platform != 'win32'
triton-windows ; sys_platform == 'win32'